* arm.h (CONDITIONAL_REGISTER_USAGE): Avoid signed/unsigned 	warning. 	* emit-rtl.c (gen_rtx_REG, set_mem_attributes_minus_bitpos, 	init_emit_once): Likewise. 	* flow.c (mark_regs_live_at_end, calculate_global_regs_live): 	Likewise. 	* function.c (assign_stack_temp_for_type): Likewise. 	* loop.c (loop_invariant_p): Likewise. 	* recog.c (push_operand): Likewise. 	* regclass.c (init_reg_sets_1): Likewise. 	* reload.c (update_auto_inc_notes): Likewise. 	* reload1.c (reload_as_needed, emit_input_reload_insns): Likewise. 	* stmt.c (expand_asm_operands): Likewise. 	* stor-layout.c (start_record_layout): Likewise.
cp:
	* class.c (layout_virtual_bases): Avoid signed/unsigned warning.

java:
	* jcf-write.c (generate_bytecode_insns): Avoid signed/unsigned
	warning.


git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@61389 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/recog.c
@@ -1316,7 +1316,7 @@ push_operand (op, mode)
 #ifdef STACK_GROWS_DOWNWARD
 	  || INTVAL (XEXP (XEXP (op, 1), 1)) != - (int) rounded_size
 #else
-	  || INTVAL (XEXP (XEXP (op, 1), 1)) != rounded_size
+	  || INTVAL (XEXP (XEXP (op, 1), 1)) != (int) rounded_size
 #endif
 	  )
 	return 0;
